+++ crates/mockiato-codegen/src/parse.rs
@@ -1,10 +1,13 @@
+use proc_macro2::Span;
+
 use crate::diagnostic::DiagnosticBuilder;
 use crate::result::Result;
-use proc_macro2::Span;
+
 pub(crate) mod method_decl;
 pub(crate) mod method_decl_parser;
 pub(crate) mod method_inputs;
 pub(crate) mod method_inputs_parser;
+pub(crate) mod mockable_attr;
 pub(crate) mod mockable_attr_parser;
 pub(crate) mod trait_decl;
 pub(crate) mod trait_decl_parser;
